requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
aio-pika==9.4.1
    # via -r requirements/base.in
aioboto3==13.1.1
    # via -r requirements/base.in
aiobotocore[boto3]==2.13.1
    # via aioboto3
aiofiles==24.1.0
    # via aioboto3
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via aiobotocore
aioitertools==0.12.0
    # via aiobotocore
aiormq==6.8.1
    # via aio-pika
aiosignal==1.3.1
    # via aiohttp
alembic==1.13.2
    # via -r requirements/base.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==24.2.0
    # via aiohttp
awscrt==0.20.11
    # via botocore
boto3[crt]==1.34.92
    # via
    #   -r requirements/base.in
    #   aiobotocore
botocore[crt]==1.34.131
    # via
    #   aiobotocore
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.1.7
    # via
    #   typer
    #   uvicorn
cryptography==43.0.3
    # via pyjwt
dependency-injector==4.42.0
    # via -r requirements/base.in
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.112.2
    # via -r requirements/base.in
fastapi-cli[standard]==0.0.5
    # via fastapi
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   yarl
jinja2==3.1.4
    # via fastapi
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mako==1.3.6
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
pamqp==3.3.0
    # via aiormq
pillow==10.4.0
    # via -r requirements/base.in
propcache==0.2.0
    # via yarl
psycopg[binary]==3.2.1
    # via -r requirements/base.in
psycopg-binary==3.2.1
    # via psycopg
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.9.0
    # via -r requirements/base.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.17
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.9.4
    # via typer
s3transfer==0.10.3
    # via boto3
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   dependency-injector
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.32
    # via
    #   -r requirements/base.in
    #   alembic
starlette==0.38.6
    # via fastapi
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
urllib3==2.2.3
    # via botocore
uvicorn[standard]==0.32.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
wrapt==1.16.0
    # via aiobotocore
yarl==1.17.1
    # via
    #   aio-pika
    #   aiohttp
    #   aiormq
